+++ pulsar/apps/data/stores/__init__.py
@@ -1,3 +1,3 @@
 from pulsar.utils.importer import import_module
-from .pulsarstore import *
+from .pulsards import *
 import_module('pulsar.apps.data.stores.redis')